♻️ refactor: Move each command to it's own file
